# schemas/openapi.json
{
  "components": {
    "schemas": {
      "Body_import_source_api_sources_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Clean born-digital PDF",
            "title": "File",
            "type": "string"
          },
          "rights_status": {
            "$ref": "#/components/schemas/RightsStatus"
          }
        },
        "required": [
          "file",
          "rights_status"
        ],
        "title": "Body_import_source_api_sources_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportJob": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Class"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "parser_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "progress_current": {
            "title": "Progress Current",
            "type": "integer"
          },
          "progress_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Total"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "state",
          "progress_current",
          "progress_total",
          "created_at",
          "updated_at"
        ],
        "title": "ImportJob",
        "type": "object"
      },
      "ImportResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/ImportJob"
          },
          "source": {
            "$ref": "#/components/schemas/SourceSummary"
          }
        },
        "required": [
          "source",
          "job"
        ],
        "title": "ImportResponse",
        "type": "object"
      },
      "JobState": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "retryable_failure",
          "needs_review",
          "failed"
        ],
        "title": "JobState",
        "type": "string"
      },
      "LessonPackage": {
        "properties": {
          "capability_notes": {
            "items": {
              "type": "string"
            },
            "title": "Capability Notes",
            "type": "array"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/CanonicalClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "compiler_version": {
            "title": "Compiler Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/SemanticFrame"
            },
            "title": "Frames",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "package_hash": {
            "title": "Package Hash",
            "type": "string"
          },
          "page_end": {
            "title": "Page End",
            "type": "integer"
          },
          "page_start": {
            "title": "Page Start",
            "type": "integer"
          },
          "schema_version": {
            "default": 2,
            "title": "Schema Version",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/SourceSummary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "verification_status": {
            "default": "draft",
            "enum": [
              "draft",
              "reviewed",
              "verified"
            ],
            "title": "Verification Status",
            "type": "string"
          },
          "visuals": {
            "items": {
              "$ref": "#/components/schemas/SourceVisual"
            },
            "title": "Visuals",
            "type": "array"
          }
        },
        "required": [
          "id",
          "package_hash",
          "title",
          "source",
          "page_start",
          "page_end",
          "compiler_version",
          "capability_notes",
          "claims",
          "visuals",
          "frames",
          "created_at"
        ],
        "title": "LessonPackage",
        "type": "object"
      },
...
      "ResearchEventIn": {
        "properties": {
          "event_type": {
            "enum": [
              "session_started",
              "frame_shown",
              "play",
              "pause",
              "step_forward",
              "step_back",
              "source_opened",
              "source_closed",
              "pace_changed",
              "focus_paused",
              "study_submitted",
              "session_ended"
            ],
            "title": "Event Type",
            "type": "string"
          },
          "frame_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Id"
          },
          "lesson_id": {
            "maxLength": 80,
            "minLength": 8,
            "title": "Lesson Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Payload",
            "type": "object"
          },
          "session_id": {
            "maxLength": 80,
            "minLength": 8,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "lesson_id",
          "event_type",
          "occurred_at"
        ],
        "title": "ResearchEventIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PRISM local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/events": {
      "post": {
        "operationId": "append_event_api_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Event"
      }
    },
...
    "/api/imports/{job_id}": {
      "get": {
        "operationId": "import_status_api_imports__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Status"
      }
    },
    "/api/imports/{job_id}/resume": {
      "post": {
        "operationId": "resume_import_api_imports__job_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Import"
      }
    },
    "/api/lessons/{lesson_id}": {
      "get": {
        "operationId": "lesson_api_lessons__lesson_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "title": "Lesson Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonPackage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lesson"
      }
    },
    "/api/lessons/{lesson_id}/events": {
      "get": {
        "operationId": "lesson_events_api_lessons__lesson_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "title": "Lesson Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResearchEventRecord"
                  },
                  "title": "Response Lesson Events Api Lessons  Lesson Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lesson Events"
      }
    },
    "/api/sources": {
      "get": {
        "operationId": "list_sources_api_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceSummary"
                  },
                  "title": "Response List Sources Api Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sources"
      },
      "post": {
        "operationId": "import_source_api_sources_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_source_api_sources_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Source"
      }
    },
    "/api/sources/{source_id}/cover": {
      "get": {
        "operationId": "source_cover_api_sources__source_id__cover_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Cover"
      }
    },
    "/api/sources/{source_id}/file": {
      "get": {
        "operationId": "source_file_api_sources__source_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source File"
      }
    },
    "/api/sources/{source_id}/lessons": {
      "post": {
        "operationId": "compile_source_api_sources__source_id__lessons_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileLessonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonPackage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compile Source"
      }
    },
    "/api/sources/{source_id}/readiness": {
      "get": {
        "operationId": "source_readiness_api_sources__source_id__readiness_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Start"
            }
          },
          {
            "in": "query",
            "name": "page_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceReadiness"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Readiness"
      }
    },
    "/api/sources/{source_id}/reading-state": {
      "get": {
        "operationId": "get_reading_state_api_sources__source_id__reading_state_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Reading State"
      },
      "put": {
        "operationId": "put_reading_state_api_sources__source_id__reading_state_put",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingStateUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Reading State"
      }
    },
    "/api/sources/{source_id}/search": {
      "get": {
        "operationId": "source_search_api_sources__source_id__search_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 40,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Search"
      }
    },
    "/api/sources/{source_id}/structure": {
      "get": {
        "operationId": "source_structure_api_sources__source_id__structure_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceStructure"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Structure"
      }
    },
    "/api/sources/{source_id}/visuals/{visual_id}": {
      "get": {
        "operationId": "source_visual_api_sources__source_id__visuals__visual_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "visual_id",
            "required": true,
            "schema": {
              "title": "Visual Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Visual"
      }
    }
  }
}